refactor: extract image manipulation functions to useImage
